Moved CatchSpecs to Should Specs library
