+++ Source/Machine.Specifications.Should.Specs/Machine.Specifications.Should.Specs.csproj
@@ -49,6 +49,7 @@
       <Link>Properties\VersionInfo.cs</Link>
     </Compile>
     <Compile Include="AssertionSpecs.cs" />
+    <Compile Include="CatchSpecs.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ShouldBeLikeSpecs.cs" />
     <Compile Include="Utility\Internal\ErrorMessageSpecs.cs" />
